refactor: initial type checking with `ty` (#144)
* refactor: make `ty` happy

* refactor: more `ty` happiness

* refactor: more `ty` happiness

* feat: more `ty` happiness

* refactor: move stats to pydantic models

* refactor: more `ty` happiness

* docs: update README

* refactor: clean Harbor client

* ci: auto-format from pre-commit.com hooks

* refactor: remove dangling file

* refactor: use more specific exception types

* refactor: fix FastAPI annotations

* refactor: make sonarqube happy

---------

+++ ai4papi/routers/v1/snapshots.py
@@ -9,16 +9,14 @@ The strategy for saving in Harbor is:
 
 from copy import deepcopy
 import datetime
-from typing import Tuple, Union
 import uuid
 
-from fastapi import APIRouter, Depends, HTTPException, Query
+from fastapi import APIRouter, Depends, HTTPException
 from fastapi.security import HTTPBearer
 from harborapi import HarborClient
 
-from ai4papi import auth
+from ai4papi import auth, nomad, schemas
 import ai4papi.conf as papiconf
-import ai4papi.nomad.common as nomad_common
 
 
 router = APIRouter(
@@ -38,8 +36,8 @@ if papiconf.HARBOR_USER and papiconf.HARBOR_PASS:
 else:
     client = None
 
-# Use the Nomad cluster inited in nomad.common
-Nomad = nomad_common.Nomad
+# Use the Nomad cluster inited in nomad utils
+Nomad = nomad.Nomad
 
 # Define limits for snapshots size
 INDIVIDUAL_LIMIT_GB = 10
@@ -48,7 +46,7 @@ TOTAL_LIMIT_GB = 15
 
 @router.get("")
 def get_snapshots(
-    vos: Union[Tuple, None] = Query(default=None),
+    vos: schemas.VoList = None,
     authorization=Depends(security),
 ):
     """
@@ -62,16 +60,19 @@ def get_snapshots(
     auth_info = auth.get_user_info(token=authorization.credentials)
 
     # If no VOs, then retrieve jobs from all user VOs
-    # Always remove VOs that do not belong to the project
-    vos = set(vos).intersection(set(papiconf.MAIN_CONF["auth"]["VO"]))
-    if not vos:
+    if vos is None:
+        user_vos = set(papiconf.MAIN_CONF["auth"]["VO"])
+    else:
+        # Always remove VOs that do not belong to the project
+        user_vos = set(vos).intersection(set(papiconf.MAIN_CONF["auth"]["VO"]))
+    if not user_vos:
         raise HTTPException(
             status_code=401,
             detail=f"Your VOs do not match available VOs: {papiconf.MAIN_CONF['auth']['VO']}.",
         )
 
     snapshots = []
-    for vo in vos:
+    for vo in user_vos:
         # Retrieve the completed snapshots from Harbor
         snapshots += get_harbor_snapshots(owner=auth_info["id"], vo=vo)
 
@@ -114,10 +115,10 @@ def create_snapshot(
         )
 
     # Load module configuration
-    nomad_conf = deepcopy(papiconf.SNAPSHOTS["nomad"])
+    nomad_template = deepcopy(papiconf.SNAPSHOTS["nomad"])
 
     # Get target job info
-    job_info = nomad_common.get_deployment(
+    job_info = nomad.get_deployment(
         deployment_uuid=deployment_uuid,
         namespace=namespace,
         owner=auth_info["id"],
@@ -134,7 +135,7 @@ def create_snapshot(
 
     # Replace the Nomad job template
     now = datetime.datetime.now()
-    nomad_conf = nomad_conf.safe_substitute(
+    nomad_conf_str = nomad_template.safe_substitute(
         {
             "JOB_UUID": uuid.uuid1(),
             "NAMESPACE": papiconf.MAIN_CONF["nomad"]["namespaces"][vo],
@@ -156,10 +157,10 @@ def create_snapshot(
     )
 
     # Convert template to Nomad conf
-    nomad_conf = nomad_common.load_job_conf(nomad_conf)
+    nomad_conf = nomad.load_job_conf(nomad_conf_str)
 
     # Submit job
-    _ = nomad_common.create_deployment(nomad_conf)
+    _ = nomad.create_deployment(nomad_conf)
 
     return {
         "status": "success",
@@ -187,7 +188,7 @@ def delete_snapshot(
     # Check is the snapshot is in the "completed" list (Harbor)
     snapshots = get_harbor_snapshots(owner=auth_info["id"], vo=vo)
     snapshot_ids = [s["snapshot_ID"] for s in snapshots]
-    if snapshot_uuid in snapshot_ids:
+    if client and (snapshot_uuid in snapshot_ids):
         _ = client.delete_artifact(
             project_name="user-snapshots",
             repository_name=auth_info["id"].replace("@", "_at_"),
@@ -226,8 +227,10 @@ def get_harbor_snapshots(
     * **vo**: Virtual Organization the snapshot belongs to
     """
     # Check if the user exists in Harbor (ie. Docker image exists)
+    if not client:
+        return []
     repos = client.get_repositories(project_name="user-snapshots")
-    users = [r.name.split("/")[1] for r in repos]
+    users = [r.name.split("/")[1] for r in repos]  # ty: ignore[not-iterable]
     user_str = owner.replace("@", "_at_")
     if user_str not in users:
         return []
@@ -238,7 +241,7 @@ def get_harbor_snapshots(
         repository_name=user_str,
     )
     snapshots = []
-    for a in artifacts:
+    for a in artifacts:  # ty: ignore[not-iterable]
         # Ignore snapshot if it doesn't belong to the VO
         a_labels = a.extra_attrs.root["config"]["Labels"]
         if a_labels.get("VO") != vo:
